fix: equality should be deep
